% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{.make_plot}
\alias{.make_plot}
\title{Make a faceted preprocessing plot per spectra and peptide}
\usage{
.make_plot(
  sp,
  peaks,
  peptides_user,
  n_isopeaks,
  norm_func = NULL,
  peptide_labeller
)
}
\arguments{
\item{n_isopeaks}{}
}
\description{
Make a faceted preprocessing plot per spectra and peptide
}
