% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{.preprocess_spectra}
\alias{.preprocess_spectra}
\title{Internal preprocessing function}
\usage{
.preprocess_spectra(
  sps_mzr,
  pep_table,
  verbose,
  parparam,
  chunk_size,
  smooth_wma_hws = 4L,
  smooth_sg_hws = 6L,
  iterations = 50L,
  halfWindowSize = 20L,
  snr = 2,
  k = 0L,
  threshold = 0.33,
  local_bg = FALSE,
  mass_range = 100,
  bg_cutoff = 0.5,
  l_cutoff = 1e-08,
  tolerance = 0.4,
  ppm = 50,
  n_isopeaks = 5,
  min_isopeaks = 4,
  norm_func = NULL
)
}
\description{
Preprocess a Spectra object with MsBackendMzR
}
