% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{.preprocessing_plot}
\alias{.preprocessing_plot}
\title{Preprocessing function for plotting}
\usage{
.preprocessing_plot(
  sps_mzr,
  pep_table,
  verbose,
  smooth_wma_hws = 4,
  smooth_sg_hws = 6,
  iterations = 50,
  halfWindowSize = 20,
  snr = 2,
  k = 0L,
  threshold = 0.33,
  local_bg = FALSE,
  mass_range = 100,
  bg_cutoff = 0.5,
  l_cutoff = 1e-08,
  tolerance = 0.4,
  ppm = 50,
  n_isopeaks = 5,
  min_isopeaks = 4,
  norm_func = NULL,
  q2e = NULL,
  peptide_labeller = NULL
)
}
\arguments{
\item{q2e}{\code{numeric}.
A theoretical isotopic invelope with this \code{q2e} is overlaid in the plot in blue.}

\item{peptide_labeller}{\code{function}.
A function used to transform the facet labels. See \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and \code{\link[ggplot2:labeller]{ggplot2::labeller()}}}
}
\description{
Preprocess a Spectra object so that it can be plotted
}
