% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{tol}{numreric tolerance. Default is \code{.Machine$double.eps^0.5}}
}
\value{
FALSE if \code{x} is not a whole number, TRUE if it is, given a numberic tolerance
}
\description{
Check whole number
}
