% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_q2e.R
\name{lm_q2e_oneshot}
\alias{lm_q2e_oneshot}
\title{Estimation of peptide q2e in a single linear regression with nested coefficients}
\usage{
lm_q2e_oneshot(
  peaks_df,
  intensity = "norm_int",
  weights = NULL,
  intercept = FALSE
)
}
\arguments{
\item{peaks_df}{Data.frame with peaks data and theoretical isotopic envelopes}

\item{intensity}{Term used as the dependent variable. It could be the unscaled
intensity or the normalized at the spectra and peptide level}

\item{weights}{Column of weights in \code{peaks_df}}

\item{intercept}{Logical, whether to fit an intercept or not}
}
\value{
A data.frame with q2e estimates
}
\description{
Use a single call to "lm" with nested coefficients per spectra and peptide.
Effectively it will calculate separate coefficient for each spectra and peptide.
}
\details{
WARNING: this cannot handle too large data.frames.
I will try to add \code{biglm} at some point.
Otherwise use group_by and the other functions
}
