% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mds.R
\name{lme_mds}
\alias{lme_mds}
\title{Calculate MDS using linear mixed effect model from q2e estimates}
\usage{
lme_mds(q2e_vals, logq = TRUE, g = NULL, outdir = NULL, return_model = F)
}
\arguments{
\item{q2e_vals}{data.frame with q2e estimations per sample, replicate and peptide}

\item{logq}{Whether q values are log-scaled before entering the LME model}

\item{g}{Reliability power in the error normal distribution from linear mixed effects model}

\item{outdir}{Optional. directory where results tables and plots are saved}

\item{return_model}{\code{logical}. Whether to return the lme model object}
}
\value{
list contaning the following:
\itemize{
\item \strong{sample}: data.frame with aggregated MDS estimates per sample
\item \strong{pep}: data.frame of fitted and residuals per replicate and peptide
\item \strong{estimates}: list of model estimates
}
\code{sample} data.frame has the following columns:
\itemize{
\item Sample: sample name
\item Prediction: sample random effects obtained from custom formula that. It is the log(MDS)
\item sd: log(MDS) standard error
\item RanefModel: sample random effects, as calculated by \code{\link[nlme]{ranef}}.
It should be the same as Prediction, up to numerical precision
\item MDS.Model, exponentials estimates. It is the MDS

}

\code{pep} data.frame contains:
\itemize{
\item Sample, replicate and peptide IDs
\item q: is the q calculated by the WLS from the isotopic distributions
\item Reliability: minimum least square of q caluclated in the WLS step
\item Loqq: log(q)
\item resp: either q or log(q), according to whether logq was FALSE or TRUE
\item Fitted: fitted q value
\item Res: pearson residuals from Fitted
\item Fitted0: fitted q values at the peptide level. It's equal to the peptide fixed effect
\item Res0: pearson residuals from Fitted0
}

\code{estimates} list contains
\itemize{
\item alpha: peptide fixed effects
\item sigma2_S: sample random effect variance
\item sigma2_R: replicate random effect variance
\item gamma: Reliability 2·gamma exponent
\item sigma2: peptide fixed effects variance
}
}
\description{
Calculate MDS using linear mixed effect model from q2e estimates
}
