% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{plot_n_peaks_per_peptide}
\alias{plot_n_peaks_per_peptide}
\title{Plotting the number of isotopic peaks detected}
\usage{
plot_n_peaks_per_peptide(peaks, n_isopeaks, min_isopeaks, marker_order, ...)
}
\arguments{
\item{marker_order}{}
}
\description{
Plotting the number of isotopic peaks detected
}
