% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{print_progress}
\alias{print_progress}
\title{Print only if verbose}
\usage{
print_progress(msg, verbose)
}
\arguments{
\item{msg}{Message to print}

\item{verbose}{\code{logical} whether to print the message}
}
\description{
Print only if verbose
}
