% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_q2e.R
\name{wls_q2e}
\alias{wls_q2e}
\title{Estimate Q2E per peptide using a variant of a least squares regression}
\usage{
wls_q2e(
  peaks,
  lm_use = c(wlm_q2e_intercept, lm_q2e_intercept, wlm_q2e, lm_q2e),
  by_spectrum = TRUE
)
}
\arguments{
\item{peaks}{data.frame. Isotopic peaks data produced by \code{\link[=preprocess_spectra]{preprocess_spectra()}}}

\item{lm_use}{Function. Model variant to use: \code{\link[=wlm_q2e_intercept]{wlm_q2e_intercept()}} (default), \link{lm_q2e_intercept},
\code{\link[=wlm_q2e]{wlm_q2e()}}, or \code{\link[=lm_q2e]{lm_q2e()}}}

\item{by_spectrum}{Logical. Whether to calculate a value per replicate/spectrum (\code{TRUE}, default)
or sample (\code{FALSE}).}
}
\value{
A \code{data.frame} contaning Q2E estimates with the following columns:
\itemize{
\item \code{sample_name}
\item \code{spectrumId} optional, only if \code{by_spectrum} is \code{TRUE}
\item \code{pep_number}
\item \code{gamma_0} coefficient of the native, non-deamidated peptide
\item \code{gamma_1} coefficient of the peptide with 1 deamidation
\item \code{gamma_2} coefficient of the peptide with 2 demidations. Only for markers with
2 glutamines, other with it is \code{NA}
\item \code{q2e} Ratio of deamidated to total. This is \deqn{1-\frac{gamma_0}{gamma_0 + gamma_1 + gamma_2}}
\item \code{intercept} optional, only if using a linear model with free intercept
\item \code{residual} Squared root of the sum of the squared residuals
\item \code{reliability} 1 - \code{residual}
}
}
\description{
Estimate Q2E per peptide using a variant of a least squares regression
}
