% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{.make_plot}
\alias{.make_plot}
\title{Make a faceted preprocessing plot per spectra and peptide}
\usage{
.make_plot(sp, peaks, peptide_labeller = NULL)
}
\arguments{
\item{peptide_labeller}{}
}
\description{
Make a faceted preprocessing plot per spectra and peptide
}
