% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{.make_plot}
\alias{.make_plot}
\title{Make a faceted spectral plot per spectra and peptide}
\usage{
.make_plot(sp, peaks, peptide_labeller = NULL)
}
\arguments{
\item{sp}{A \code{data.frame} with raw and preprocessed profile spectral data}

\item{peaks}{A \code{data.frame} with peaks data}

\item{peptide_labeller}{A ggplot \code{\link[ggplot2:labeller]{ggplot2::labeller()}} type function}
}
\description{
It shows the raw, processed spectra and detected peaks per peptide marker
region and spectrum.
}
