% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mds.R
\name{predict_mds}
\alias{predict_mds}
\title{Title}
\usage{
predict_mds(model, estimates, new_q2e = NULL, logq = T)
}
\arguments{
\item{model}{lme model object}

\item{estimates}{}

\item{new_q2e}{New q2e data for which MDS is predicted using the trained model
q is stored in a column called "resp"}

\item{logq}{whether q is in log-scale in q2e or new_q2e}
}
\description{
Title
}
