% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_mds.R
\name{predict_sample}
\alias{predict_sample}
\title{Perform prediction of MDS and associated SD on a given sample}
\usage{
predict_sample(sample_name, spectrumId, pep_number, reliability, resp, pars)
}
\arguments{
\item{sample_name}{Sample name}

\item{spectrumId}{Unique spectrum ID or replicate within a sample}

\item{pep_number}{Peptide identifier}

\item{reliability}{Reliability (1-residual)}

\item{resp}{Response variable, either q or Logq}

\item{pars}{Contains the following estimates as extracted by \link{extract_estimates}:
\describe{
\item{data_I_s}{part of the data frame corresponding to Sample s.}
\item{alpha}{estimate for fixed effect}
\item{sigma2_S}{estimate for variance for random effect of sample.}
\item{sigma2_R}{estimate for variance for random effect of replication.}
\item{gamma}{estimate for half power on Reliability}
\item{sigma2}{estimate for variance for on the different peptides. This should be a named vector.}
}}
}
\value{
tibble with predicted sample MDS and standard deviation
}
\description{
Perform prediction of MDS and associated SD on a given sample
}
